added the useToastMessages custom hook
